Updating Ethereum differences section (#57)
* Change title to reflect brand; clarified opcode 

It looks like this section was previously named 'Ethereum and zkEVM differences', and then 'zkEVM' got removed, but 'Linea' didn't get inserted. 

I also changed the main heading to "Differences between Ethereum and Linea', for clarity, as 'Ethereum and Linea differences' still leaves some semantic ambiguity.

I also added some additional context and explanation to the opcode section, as if you go to the EF's opcode reference and search "Difficulty", you won't find anything, which could be confusing to some. 

There is significant documentation around about how developers should compensate for this, specifically by avoiding relying on specific values for what is returned by PREVRANDAO, and instead using more abstract logic. For example, (here)[https://soliditydeveloper.com/prevrandao]. If there's a specific explainer that we feel is solid, or a good go-to, we could link to that.

* Update ethereum-differences.md

It looks like this section was previously named 'Ethereum and zkEVM differences', and then 'zkEVM' got removed, but 'Linea' didn't get inserted. 

I also changed the main heading to "Differences between Ethereum and Linea', for clarity, as 'Ethereum and Linea differences' still leaves some semantic ambiguity.

I also added some additional context and explanation to the opcode section, as if you go to the EF's opcode reference and search "Difficulty", you won't find anything, which could be confusing to some. 

There is significant documentation around about how developers should compensate for this, specifically by avoiding relying on specific values for what is returned by PREVRANDAO, and instead using more abstract logic. For example, (here)[https://soliditydeveloper.com/prevrandao]. If there's a specific explainer that we feel is solid, or a good go-to, we could link to that.

* fixing linter error: escape `RANDAO`

* Adding RANDAO 

This is in order to better document the PREVRANDAO opcode.

* added randao to the dictionary; no need to escape